[+] Add lib/targeting.rb
targeting.rb is a script that reads the RegDate header from event_attendees.csv and creates
a new #Time object intance from it. With this new 'time' object, the script prints the
entire list of date, time, and weekday of each registration entry. It also shows a hash
of the most frequent dates, times, and weekdays.